requirements.txt
antiorm==1.2.1
asgiref==3.7.2
db==0.1.1
db-sqlite3==0.0.1
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
Faker==19.13.0

